requirements.txt
httpx==0.27.0
Pillow==10.4.0
loguru==0.7.2
# 2.7.x не собирается на Python 3.13 (PyO3); >=2.10 даёт готовые wheels
pydantic>=2.10,<3
gmsaas>=1.12.0
playwright>=1.49.0
# BrowserStack App Automate (реальные Android устройства)
Appium-Python-Client>=3.1.0
coverage>=7.6.0
pytest-cov>=5.0.0
